Implement Dear ImGui w/ minor other stuff
This is looking extremely promising! Still work to do re: window size
etc, but the performance is amazing. Good enough that I literally can't
figure out what the frametime impact is; sometimes it seems like 0.1 ms,
but there are times when showing the window REDUCES the frametime, so
the variance is clearly on the order of the actual slowdown, if not
greater!