Add persistent settings!
All settings are currently made within the mod UI. There are no settings
in the Noita "Mod Settings" panel.

I think I'll keep it this way in the future since it works and it much
nicer to work with -- seeing the changes in real-time, etc.

I may add a setting to ignore all mouse input, in which case I will add
that to the Noita mod menu, since it will be impossible to revert
otherwise!