Make whitespace consistent (4 spaces)
I have no idea how I ended up with spaces in one file and tabs in the
other, but... yeah.
After moving some code between the two I ended up with a mix of tabs and
spaces.

And to add insult to injury, utils.lua had mostly 2 spaces.

git log -wp should show that this commit changes nothing whatsoever that
is not whitespace.