Fix cursed area damage showing as Holy Mountain
I don't think Noita used to report such damage as $biome_holymountain,
but regardless, it does now (in the April 30, 2024 build).

The big issue with this isn't necessarily the incorrect source, but that
it causes log spam with every other row Holy Mountain, every other row
Cursed rock.

This is an ugly fix, but I'm not sure a better one is possible on my
end.