Add "log_healing" setting
Ideally this would be a setting to show healing, which is always logged,
since that would be more user-friendly and have an unnoticeable
performance impact, but it was too much work.

Simply filtering out healing on display means that if 40 of the last 60
hits were healing and the user said to display 60 rows, only 20 would be
shown.